refactor: create new folder constant
