+++ src/main/java/app/bottlenote/alcohols/domain/constant/AlcoholType.java
@@ -1,4 +1,4 @@
-package app.bottlenote.alcohols.domain;
+package app.bottlenote.alcohols.domain.constant;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
