+++ src/main/java/app/bottlenote/rating/domain/Rating.java
@@ -37,6 +37,5 @@ public class Rating extends BaseEntity {
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id")
 	private User user;
-
 }
