--- src/main/java/app/bottlenote/review/domain/SizeType.java
@@ -1,7 +1,0 @@
-package app.bottlenote.review.domain;
-
-public enum SizeType{
-
-	GLASS, BOTTLE
-
-}
