+++ src/main/java/app/bottlenote/review/domain/constant/SizeType.java
@@ -1,0 +1,7 @@
+package app.bottlenote.review.domain.constant;
+
+public enum SizeType{
+
+	GLASS, BOTTLE
+
+}
