+++ src/main/java/app/bottlenote/support/report/domain/UserReports.java
@@ -1,6 +1,7 @@
 package app.bottlenote.support.report.domain;
 
 import app.bottlenote.common.domain.BaseEntity;
+import app.bottlenote.support.constant.StatusType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
